+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Added
 
 - Support for QHD displays
-- `sandstone/PageViews` prop `fullContents`
+- `sandstone/PageViews` prop `fullContents` to allow contents of the page to use the entire area
 
 ### Fixed